kkuhn removes another 16 yellow flags and a couple of warnings to make it 14 yellow flags no warnings

+++ PlanningPoker/src/edu/wpi/cs/wpisuitetng/modules/planningpoker/view/ToolbarView.java
@@ -72,10 +72,13 @@ public class ToolbarView  extends DefaultToolbarView {
 		buttonPanel.removeButtons();
 	}
 
-	
+	/**
+	 * gets the available width
+	 * @return the size in an integer
+	 */
 	public int getDummySize()
 	{
-		int size = (int) getWidth();
+		int size = getWidth();
 		size -= buttonPanel.getWidth();
 		size -= gameButton.getWidth();
